refactured the global chats loader

+++ src/events/globalchat/messages/delete.js
@@ -4,8 +4,6 @@ module.exports = {
 
 module.exports.call = async (bot, msg) => {
     const am = bot.cache.globalchat.get(msg.id);
-    if(am){
-        am.forEach(ms => ms.delete().catch(()=>{}));
-    }
+    if(am) am.forEach(ms => ms.delete().catch(()=>{}));
     bot.cache.globalchat.delete(msg.id);
 }
